removes orgsprovider in favor of just using tanstack directly
