--- src/features/orgs/index.d.ts
@@ -1,9 +1,0 @@
-import type { IAltinnOrgs } from 'src/types/shared';
-
-export interface IOrgsState {
-  allOrgs: IAltinnOrgs | null;
-}
-
-export interface IFetchOrgsFulfilled {
-  orgs: IAltinnOrgs;
-}
